requirements.txt
streamlit==1.37.*
gspread==6.1.*
google-auth==2.34.*
tornado==6.4.1
gspread-dataframe~=4.0.0
pandas~=2.3.1
google-api-python-client~=2.184.0
